# swagger.yaml
openapi: 3.0.3
info:
  title: shop_ease_api
  version: 1.0.0
  contact: {}
servers:
  - url: http://localhost:5000
    description: Local server
  - url: https://api.getpostman.com
    description: Postman server
paths:
...
  /api/v1/users:
    get:
      summary: getAllUsers
      description: getAllUsers
      operationId: getallusers
      responses:
        '200':
          description: ''
    patch:
      summary: updateUser
      description: updateUser
      operationId: updateuser
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: hezron4@gmail.com
                name:
                  type: string
                  example: john junior
            examples:
              updateUser:
                value:
                  email: hezron4@gmail.com
                  name: john junior
      responses:
        '200':
          description: ''
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      name: X-API-Key
      in: header
...